suez-app/requirements.txt
Flask==3.0.0
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.1.0
openpyxl==3.1.2
gunicorn==21.2.0
